+++ crates/store/src/server/api.rs
@@ -8,16 +8,16 @@ use miden_node_proto::{
         account::AccountSummary,
         note::NoteSyncRecord,
         requests::{
-            ApplyBlockRequest, CheckNullifiersRequest, GetAccountDetailsRequest,
-            GetAccountStateDeltaRequest, GetBlockByNumberRequest, GetBlockHeaderByNumberRequest,
-            GetBlockInputsRequest, GetNoteInclusionProofsRequest, GetNotesByIdRequest,
+            ApplyBlockRequest, CheckNullifiersByPrefixRequest, CheckNullifiersRequest,
+            GetAccountDetailsRequest, GetAccountStateDeltaRequest, GetBlockByNumberRequest,
+            GetBlockHeaderByNumberRequest, GetBlockInputsRequest, GetNoteInclusionProofsRequest, GetNotesByIdRequest,
             GetTransactionInputsRequest, ListAccountsRequest, ListNotesRequest,
             ListNullifiersRequest, SyncStateRequest,
         },
         responses::{
-            AccountTransactionInputRecord, ApplyBlockResponse, CheckNullifiersResponse,
-            GetAccountDetailsResponse, GetAccountStateDeltaResponse, GetBlockByNumberResponse,
-            GetBlockHeaderByNumberResponse, GetBlockInputsResponse, GetNoteInclusionProofsResponse,
+            AccountTransactionInputRecord, ApplyBlockResponse, CheckNullifiersByPrefixResponse,
+            CheckNullifiersResponse, GetAccountDetailsResponse, GetAccountStateDeltaResponse,
+            GetBlockByNumberResponse, GetBlockHeaderByNumberResponse, GetBlockInputsResponse,
             GetNotesByIdResponse, GetTransactionInputsResponse, ListAccountsResponse,
             ListNotesResponse, ListNullifiersResponse, NullifierTransactionInputRecord,
             NullifierUpdate, SyncStateResponse,
@@ -111,6 +111,41 @@ impl api_server::Api for StoreApi {
         Ok(Response::new(CheckNullifiersResponse { proofs: convert(proofs) }))
     }
 
+    /// Returns nullifiers that match the specified prefixes and have been consumed.
+    ///
+    /// Currently the only supported prefix length is 16 bits.
+    #[instrument(
+        target = "miden-store",
+        name = "store:check_nullifiers_by_prefix",
+        skip_all,
+        ret(level = "debug"),
+        err
+    )]
+    async fn check_nullifiers_by_prefix(
+        &self,
+        request: tonic::Request<CheckNullifiersByPrefixRequest>,
+    ) -> Result<Response<CheckNullifiersByPrefixResponse>, Status> {
+        let request = request.into_inner();
+
+        if request.prefix_len != 16 {
+            return Err(Status::invalid_argument("Only 16-bit prefixes are supported"));
+        }
+
+        let nullifiers = self
+            .state
+            .check_nullifiers_by_prefix(request.prefix_len, request.nullifiers)
+            .await
+            .map_err(internal_error)?
+            .into_iter()
+            .map(|nullifier_info| NullifierUpdate {
+                nullifier: Some(nullifier_info.nullifier.into()),
+                block_num: nullifier_info.block_num,
+            })
+            .collect();
+
+        Ok(Response::new(CheckNullifiersByPrefixResponse { nullifiers }))
+    }
+
     /// Returns info which can be used by the client to sync up to the latest state of the chain
     /// for the objects the client is interested in.
     #[instrument(
@@ -158,7 +193,7 @@ impl api_server::Api for StoreApi {
             .notes
             .into_iter()
             .map(|note| NoteSyncRecord {
-                note_index: note.note_index.to_absolute_index() as u32,
+                note_index: note.note_index.to_absolute_index(),
                 note_id: Some(note.note_id.into()),
                 metadata: Some(note.metadata.into()),
                 merkle_path: Some(Into::into(&note.merkle_path)),